LayoutTests/imported/w3c/web-platform-tests/loading/early-hints/referrer-policy-origin-when-cross-origin.h2.window-expected.txt

CONSOLE MESSAGE: <link rel=preload> has an invalid `href` value

FAIL Referrer policy: origin-when-cross-origin assert_true: expected true got false
